# core, all: split vm.Context into BlockContext and TxContext