As you can see, this is a mixed bag, ring tx are way faster, while other benchmarks are a bit slower

This PR:
```
BenchmarkInsertChain_empty_memdb-8            	    4575	    306455 ns/op	  178274 B/op	     510 allocs/op
BenchmarkInsertChain_empty_diskdb-8           	    3570	    351276 ns/op	  182059 B/op	     532 allocs/op
BenchmarkInsertChain_valueTx_memdb-8          	    3625	    373037 ns/op	  203530 B/op	     684 allocs/op
BenchmarkInsertChain_valueTx_diskdb-8         	    3140	    376081 ns/op	  207843 B/op	     705 allocs/op
BenchmarkInsertChain_valueTx_100kB_memdb-8    	    1000	   1521589 ns/op	  746068 B/op	     684 allocs/op
BenchmarkInsertChain_valueTx_100kB_diskdb-8   	     758	   1472932 ns/op	 1056727 B/op	     709 allocs/op
BenchmarkInsertChain_uncles_memdb-8           	    3351	    498745 ns/op	  195669 B/op	     689 allocs/op
BenchmarkInsertChain_uncles_diskdb-8          	    2412	    489038 ns/op	  201543 B/op	     707 allocs/op
BenchmarkInsertChain_ring200_memdb-8          	     146	   8301682 ns/op	 4108689 B/op	   30131 allocs/op
BenchmarkInsertChain_ring200_diskdb-8         	     144	   8333911 ns/op	 4153333 B/op	   29473 allocs/op
BenchmarkInsertChain_ring1000_memdb-8         	     120	   9980914 ns/op	 4597596 B/op	   34056 allocs/op
BenchmarkInsertChain_ring1000_diskdb-8        	     100	  10021250 ns/op	 4659023 B/op	   33422 allocs/op
```
Master:
```
BenchmarkInsertChain_empty_memdb-8            	    5017	    217500 ns/op	  177494 B/op	     505 allocs/op
BenchmarkInsertChain_empty_diskdb-8           	    5956	    221143 ns/op	  180055 B/op	     524 allocs/op
BenchmarkInsertChain_valueTx_memdb-8          	    5241	    274443 ns/op	  203068 B/op	     699 allocs/op
BenchmarkInsertChain_valueTx_diskdb-8         	    5948	    271572 ns/op	  206797 B/op	     717 allocs/op
BenchmarkInsertChain_valueTx_100kB_memdb-8    	    1767	   1085632 ns/op	  741137 B/op	     698 allocs/op
BenchmarkInsertChain_valueTx_100kB_diskdb-8   	    1005	   1071035 ns/op	 1064578 B/op	     724 allocs/op
BenchmarkInsertChain_uncles_memdb-8           	    3931	    354559 ns/op	  194458 B/op	     683 allocs/op
BenchmarkInsertChain_uncles_diskdb-8          	    4341	    385686 ns/op	  199261 B/op	     703 allocs/op
BenchmarkInsertChain_ring200_memdb-8          	     120	  10077005 ns/op	 4594716 B/op	   35317 allocs/op
BenchmarkInsertChain_ring200_diskdb-8         	      97	  11685285 ns/op	 4646687 B/op	   34663 allocs/op
BenchmarkInsertChain_ring1000_memdb-8         	      86	  13789912 ns/op	 5194921 B/op	   39543 allocs/op
BenchmarkInsertChain_ring1000_diskdb-8        	      94	  14345609 ns/op	 5252341 B/op	   38873 allocs/op
```