@MariusVanDerWijden 2020-10-13
Adding these changes on top of the passing vm.Context by pointer produces the following results:
```

BenchmarkInsertChain_empty_memdb-8            	    6682	    184174 ns/op	  177967 B/op	     518 allocs/op
BenchmarkInsertChain_empty_diskdb-8           	    7000	    227932 ns/op	  180654 B/op	     537 allocs/op
BenchmarkInsertChain_valueTx_memdb-8          	    6634	    209066 ns/op	  190173 B/op	     697 allocs/op
BenchmarkInsertChain_valueTx_diskdb-8         	    6020	    237992 ns/op	  193433 B/op	     715 allocs/op
BenchmarkInsertChain_valueTx_100kB_memdb-8    	    1734	    785544 ns/op	  728628 B/op	     697 allocs/op
BenchmarkInsertChain_valueTx_100kB_diskdb-8   	    1501	   1020587 ns/op	 1041337 B/op	     734 allocs/op
BenchmarkInsertChain_uncles_memdb-8           	    5558	    272765 ns/op	  193894 B/op	     697 allocs/op
BenchmarkInsertChain_uncles_diskdb-8          	    5248	    315847 ns/op	  199513 B/op	     716 allocs/op
BenchmarkInsertChain_ring200_memdb-8          	     183	   6714158 ns/op	 2987597 B/op	   32318 allocs/op
BenchmarkInsertChain_ring200_diskdb-8         	     177	   6954004 ns/op	 3048000 B/op	   31668 allocs/op
BenchmarkInsertChain_ring1000_memdb-8         	     147	   8271707 ns/op	 3637109 B/op	   36795 allocs/op
BenchmarkInsertChain_ring1000_diskdb-8        	     142	   8579421 ns/op	 3729780 B/op	   36126 allocs/op
```