@holiman 2020-10-15
![Screenshot_2020-10-15 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/96084603-a550a400-0ebf-11eb-8681-81c4e7a588f0.png)
This is taken after a fast-sync, about 4 hours after both machines reached the head. This PR running on `bench03` (right), against master on `bench04` (left). The `execution` appears to be a few milliseconds faster on this PR. 
It might be a good idea to swap the machines in a while, and see if the same effect it still seen.