@holiman 2020-10-18
Here's 3.5 days of block processing. This PR on the right-hand side

![Screenshot_2020-10-18 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/96374717-ae1dd000-1174-11eb-9d89-7b789c5a9692.png)

Zooming in on `execution`
![Screenshot_2020-10-18 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/96374720-b544de00-1174-11eb-9f62-f9d118559eb7.png)