@holiman 2020-10-19
Last six hours with them swapped: 

![Screenshot_2020-10-19 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/96431230-86297d80-1203-11eb-8f00-9aea4fff40e6.png)
![Screenshot_2020-10-19 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/96431243-89bd0480-1203-11eb-8a4d-16b77ec75267.png)