@holiman 2020-10-19
PR on `bench03`, master on `bench04`: `41.5ms` vs `45.1ms` == PR is `8%` faster on `execution`
PR on `bench04`, mater on `bench03`: `50.2ms` vs `56.1ms` == PR is `11%` faster on `execution`. 
So the 'real' number might be somwhere around 9.5% faster execution, shaving off `~5ms` in total block processing, out of `200ms` (overall `2.5%` faster )